requirements.txt
pytest
pytest-selenium
termcolor
allure-pytest
